vendor_management_system/requirements.txt
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
